Time: 285 ms (8.62%), Space: 36.8 MB (74.47%) - LeetHub

+++ 1383-maximum-performance-of-a-team/1383-maximum-performance-of-a-team.cpp
@@ -26,6 +26,6 @@ public:
             ans = max(ans,v[i].first*total_sum);
         }
         
-        return ans%(1000000000+7);
+        return (int)(ans%(1000000000+7));
     }
 };
